pyproject.toml
[project]
name = "voi-agente-gpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langgraph>=1.1.2",
    "pyttsx3>=2.99",
    "scipy>=1.17.1",
    "sounddevice>=0.5.5",
    "websocket-client>=1.9.0",
]
